Added better physics and steering
This commit adds better steering and physics.

Steering now exists. You can now update the steering moving the car around, left or right.

The physics are also better friction drag and everything now exist.

Now the car's parts itself are stored in a config class file where it is completely decoupled from both of the physics classes. There's also a speed limit now to each car.

Also made it so the controller is the main talking point to the car and physics.